challenges

+++ backend/src/routes/hotelsRouter.routes.js
@@ -4,9 +4,13 @@ import multer from 'multer';
 import uploadConfig from '../config/upload';
 
 import HotelController from '../controllers/HotelController';
+import UserController from '../controllers/UserController';
 import Hotel from '../models/Hotel';
+import User from '../models/User';
+import AppError from '../errors/AppError';
 
 const hotelController = new HotelController(Hotel);
+const userController = new UserController(User);
 
 const hotelsRouter = new Router();
 
@@ -27,10 +31,17 @@ hotelsRouter.get('/:name', async (req, res) => {
 });
 
 hotelsRouter.post('/', upload.single('image'), async (req, res) => {
+  const { id } = req.headers;
   const {
     name, uf, city, street, quantity, dailyValue,
   } = req.body;
   const { image } = req.file;
+
+  const user = await userController.show(id);
+
+  if (!user.profile === 'admin') {
+    throw new AppError(401, "Only admin user's can do this operation.");
+  }
 
   const hotel = await hotelController.store({
     name, uf, city, street, quantity, dailyValue, image,
@@ -44,6 +55,12 @@ hotelsRouter.put('/:id', async (req, res) => {
   const {
     name, uf, city, street, quantityRoom, dailyValue,
   } = req.body;
+
+  const user = await userController.show(id);
+
+  if (!user.profile === 'admin') {
+    throw new AppError(401, "Only admin user's can do this operation.");
+  }
 
   const hotel = await hotelController.update({
     id,
@@ -60,6 +77,12 @@ hotelsRouter.put('/:id', async (req, res) => {
 
 hotelsRouter.delete('/:id', async (req, res) => {
   const { id } = req.params;
+
+  const user = await userController.show(id);
+
+  if (!user.profile === 'admin') {
+    throw new AppError(401, "Only admin user's can do this operation.");
+  }
 
   await hotelController.delete({ id });
